EDU-211 Move api clients to dedicated folder
