+++ src/components/pages/docs.js
@@ -18,8 +18,8 @@ import { getGlobalAlerts } from '../../ui/global-alerts.js';
 import LanguageSelect from '../localization/language-select.js';
 import { Form, Input, Modal, Table, Button, Switch } from 'antd';
 import { confirmDocumentDelete } from '../confirmation-dialogs.js';
-import DocumentApiClient from '../../services/document-api-client.js';
 import LanguageNameProvider from '../../data/language-name-provider.js';
+import DocumentApiClient from '../../api-clients/document-api-client.js';
 import CountryFlagAndName from '../localization/country-flag-and-name.js';
 import permissions, { hasUserPermission } from '../../domain/permissions.js';
 import { documentMetadataShape, translationProps, languageProps, pageNameProps } from '../../ui/default-prop-types.js';
