+++ src/components/room-creation-modal.js
@@ -6,8 +6,8 @@ import { useTranslation } from 'react-i18next';
 import errorHelper from '../ui/error-helper.js';
 import { Form, Modal, Input, Radio } from 'antd';
 import { useService } from './container-context.js';
-import RoomApiClient from '../services/room-api-client.js';
 import { ROOM_ACCESS_LEVEL } from '../domain/constants.js';
+import RoomApiClient from '../api-clients/room-api-client.js';
 import React, { useState, useRef, useEffect, useCallback } from 'react';
 
 const FormItem = Form.Item;
