+++ src/services/document-import-task-processor.spec.js
@@ -2,8 +2,8 @@ import sinon from 'sinon';
 import Cdn from '../repositories/cdn.js';
 import Database from '../stores/database.js';
 import uniqueId from '../utils/unique-id.js';
-import ExportApiClient from './export-api-client.js';
 import ServerConfig from '../bootstrap/server-config.js';
+import ExportApiClient from '../api-clients/export-api-client.js';
 import DocumentImportTaskProcessor from './document-import-task-processor.js';
 import { SOURCE_TYPE as VIDEO_SOURCE_TYPE } from '../plugins/video/constants.js';
 import { destroyTestEnvironment, pruneTestEnvironment, setupTestEnvironment } from '../test-helper.js';
